change from custom JWT auth to spring security's oauth2 resource server
Remove the custom JWT auth implementation that uses custom filters, and
use spring security's OAuth2 resource server with JWT tokens, thanks to
Dan Vega. This is simpler and feels more correct as it utilizes spring
security as it is intended. Refactor auth controller/service to apply
the new changes and reduce code. Add public key and private key
generation on startup for JWT.