refactor book cleanup logic (enum) and better OOP practices
Change book delete logic by returning an enum when a book is deleted
that will say if the book's shelves list is empty, removed, or error.
Move add book/shelf logic to their entity class, better OOP practice to
keep the logic in the entity instead of repeating it in service class.